BIG updates to teams
Created register_team and add_player_to_team in teammanagerLL. created the Team class in models. Created load_teams, save_teams, add_team in the teamrepo. Made a simple test UI for creating a team, and for adding a player to a team. Einar can maybe take that and add it to the real UI.